Move attack event rate limiter to own file
